Flask_Backend/requirements.txt
Flask==2.2.3
Flask-Cors==4.0.0
Werkzeug==2.2.2
mysql-connector-python==8.3.0
PyJWT==2.8.0